requirements.txt
adal==1.2.1
applicationinsights==0.11.9
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
azure-common==1.1.21
azure-graphrbac==0.61.0
azure-mgmt-authorization==0.52.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==2.2.0
azure-mgmt-storage==3.3.0
azure-nspkg==3.0.2
azureml-automl-core==1.0.39.1
azureml-contrib-notebook==1.0.39
azureml-core==1.0.39
azureml-dataprep==1.1.3
azureml-dataprep-native==12.0.1
azureml-pipeline==1.0.39
azureml-pipeline-core==1.0.39
azureml-pipeline-steps==1.0.39
azureml-sdk==1.0.39
azureml-telemetry==1.0.39
azureml-train==1.0.39
azureml-train-automl==1.0.39.1
azureml-train-core==1.0.39
azureml-train-restclients-hyperdrive==1.0.39
azureml-widgets==1.0.39
backcall==0.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.6
bleach==3.1.0
boto==2.49.0
boto3==1.9.155
botocore==1.12.155
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
configparser==3.7.4
contextlib2==0.5.5
cryptography==2.6.1
databricks-cli==0.8.6
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
distro==1.4.0
docker==4.0.1
docutils==0.14
dotnetcore2==2.1.8
entrypoints==0.3
flake8==3.7.7
Flask==1.0.3
gensim==3.7.3
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.13.3
jeepney==0.4
Jinja2==2.10.1
jmespath==0.9.4
JsonForm==0.0.2
jsonpickle==1.2
jsonschema==3.0.1
JsonSir==0.0.2
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras2onnx==1.4.0
lightgbm==2.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mleap==0.8.1
mlflow==0.9.1
msrest==0.6.6
msrestazure==0.6.0
nbconvert==5.5.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
nimbusml==0.6.5
nose==1.3.7
nose-exclude==0.5.0
notebook==5.7.8
numpy==1.16.2
oauthlib==3.0.1
onnx==1.5.0
onnxconverter-common==1.4.2
onnxmltools==1.4.0
pandas==0.23.4
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.4.0
pathspec==0.5.9
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.2
python-dateutil==2.8.0
Python-EasyConfig==0.1.7
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.5.0
querystring-parser==1.2.3
requests==2.22.0
requests-oauthlib==1.2.0
Resource==0.2.1
ruamel.yaml==0.15.89
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.1.0
SecretStorage==3.1.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
skl2onnx==1.4.7
sklearn-pandas==1.7.0
smart-open==1.8.3
smmap2==2.0.5
sqlparse==0.3.0
statsmodels==0.9.0
tabulate==0.8.3
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.7.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.4
widgetsnbextension==3.4.2
